Add seed data

+++ Core/Entities/Member.cs
@@ -4,8 +4,8 @@ namespace Core.Entities
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Points { get; set; }
-        public int VND { get; set; }
+        public float Points { get; set; }
+        public decimal VND { get; set; }
 
     }
 }
